AppTestRunner 도입해서, 중복제거
