+++ src/test/java/AppTests.java
@@ -1,3 +1,4 @@
+import com.ll.App;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -31,4 +32,79 @@ public class AppTests {
 
         assertThat(rs).isEqualTo("안녕");
     }
+    // 테스트유틸 테스트 끝
+
+    // 앱 테스트 시작
+    @Test
+    @DisplayName("프로그램 시작시 타이틀 출력 그리고 종료")
+    public void t3() {
+        Scanner sc = TestUtil.genScanner("종료");
+        ByteArrayOutputStream output = TestUtil.setOutToByteArray();
+
+        new App(sc).run();
+
+        String rs = output.toString();
+        TestUtil.clearSetOutToByteArray(output);
+
+        assertThat(rs)
+                .contains("== 명언 앱 ==")
+                .contains("명령) ")
+                .contains("프로그램이 종료되었습니다.")
+                .doesNotContain("올바르지 않은 명령입니다.");
+    }
+
+    @Test
+    @DisplayName("잘못된 명령어 입력에 대한 처리")
+    public void t4() {
+        Scanner sc = TestUtil.genScanner("""
+                종료2
+                종료
+                """.stripIndent().trim());
+        ByteArrayOutputStream output = TestUtil.setOutToByteArray();
+
+        new App(sc).run();
+
+        String rs = output.toString();
+        TestUtil.clearSetOutToByteArray(output);
+
+        assertThat(rs)
+                .contains("올바르지 않은 명령입니다.");
+    }
+    @Test
+    @DisplayName("등록화면에서 명언과 작가를 입력받고, 명언을 생성한다.")
+    public void t5() {
+        String rs = AppTestRunner.run("""
+                등록
+                현재를 사랑하라.
+                작자미상
+                """);
+
+        assertThat(rs)
+                .contains("명언 : ")
+                .contains("작가 : ")
+                .contains("1번 명언이 등록되었습니다.");
+    }
+
+    @Test
+    @DisplayName("명언이 등록될 때 마다 생성되는 명언의 번호가 1씩 증가한다.")
+    public void t6() {
+        String rs = AppTestRunner.run("""
+                등록
+                현재를 사랑하라.
+                작자미상
+                등록
+                나의 죽음을 적들에게 알리지 마라.
+                이순신
+                등록
+                왼손은 거들뿐
+                강백호
+                """);
+
+        assertThat(rs)
+                .contains("1번 명언이 등록되었습니다.")
+                .contains("2번 명언이 등록되었습니다.")
+                .contains("3번 명언이 등록되었습니다.")
+                .doesNotContain("4번 명언이 등록되었습니다.");
+    }
+    // 앱 테스트 끝
 }
