+++ src/test/java/TestUtil.java
@@ -29,43 +29,4 @@ public class TestUtil {
             throw new RuntimeException(e);
         }
     }
-    // 테스트유틸 테스트 끝
-
-    // 앱 테스트 시작
-    @Test
-    @DisplayName("프로그램 시작시 타이틀 출력 그리고 종료")
-    public void t3() {
-        Scanner sc = TestUtil.genScanner("종료");
-        ByteArrayOutputStream output = TestUtil.setOutToByteArray();
-
-        new App(sc).run();
-
-        String rs = output.toString();
-        TestUtil.clearSetOutToByteArray(output);
-
-        assertThat(rs)
-                .contains("== 명언 앱 ==")
-                .contains("명령) ")
-                .contains("프로그램이 종료되었습니다.")
-                .doesNotContain("올바르지 않은 명령입니다.");
-    }
-
-    @Test
-    @DisplayName("잘못된 명령어 입력에 대한 처리")
-    public void t4() {
-        Scanner sc = TestUtil.genScanner("""
-                종료2
-                종료
-                """.stripIndent().trim());
-        ByteArrayOutputStream output = TestUtil.setOutToByteArray();
-
-        new App(sc).run();
-
-        String rs = output.toString();
-        TestUtil.clearSetOutToByteArray(output);
-
-        assertThat(rs)
-                .contains("올바르지 않은 명령입니다.");
-    }
-    // 앱 테스트 끝
 }
